model.c: Integrate flux band RT model
Introduce a dedicated band_flux function which returns the energy
which is transferred up and down at each level.
Use this function in order to model an atmosphere with three windows,
thereby getting rid of heating and cooling functions while preserving
the effects of both.

Pretty print the final results.