model.c: Retrieve starting temperature from table
Read in the temperature with which to start from the US standard
atmosphere table. This makes the relative humidity levels introduced in
c6468cfff0 reasonable.

In order to properly estimate temperatures for more than 20 layers,
interpolate the results in between. Hereby, generalize the concentration
interpolation function, add a new one and change its name and prototype.