model.c: Incorporate latent heat capacity
Incorporate the latent heat capacity by adapting the normal heat
capacity.
Tune the model again afterwards as to make the model converge to
sensible values.

Furthermore, split the ground albedo for short wavelength and long
wavelength calculations into separate variables. Set the long wavelength
albedo to zero as it is not fully honored in the calculations and can be
modelled using the short wavelength one only.
This fixes a conceptual mistake made in eb6e74a640.