model.c: Re-parameterize model after faulty albedo
The short-wavelength albedo was specified to be zero due to an error in
its definition using integer division. Fix this and re-parameterize the
model, i.e., the clouds.

Furthermore, introduce some dots after integers which are supposed to be
doubles at other places as well. Although, they did not introduce any
errors as of now, it is clearer this way.